checkScreenshot: Implement 'Stas Gradient Compare' algorithm based upon Skiko
The previous algorithm handled Gradients by using a hand-tuned
penalty score. However, the global 'threshold' still might
be hard to understand and could lead to strange edge cases.

The new algorithm is simpler:
Within a given area (radius), we're looking for reasonable pixel values nearby:

If we find a pixel equal or lower and one equal or higher in the area, then we can assume the pixel value to be within a gradient.

Thanks to 'stanislav.erokhin' for the algorithm.