Orchestration: Lenient packageType & opaque orchestration messages
Lenient packageType:
Previously, the OrchestrationIO would fail when a package with an unknown package type was received.
This might happen if a new package type is introduced and an old client/server sees this package type.
In order to simplify keeping forward compatibility, the IO will just skip packages of unknown type.

Opaque Orchestration Messages:
Servers which might not be able to deserialize a provided message (e.g. because of
this message being introduced in a newer version of the protocol) are supposed
to still dispatch this message to clients. This shall also help in providing forward compatibility.
Any such message is considered 'opaque' to the server (aka the server does not know what message this is).
Such opaque messages will still be emitted and distributed to all connected clients.